add list of known serial adapters used on dev boards
serial ports from these adapters are used by default when autodetecting serial port
